src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Rashid <rdmingazov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]
isort = "\"5.6.4\""
black = "\"19.10b0\""
flake8-polyfill = "\"1.0.2\""
flake8 = "\"^4.0.1\""

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
